pyproject.toml
[project]
name = "wireml"
version = "0.2.0"
description = "Teachable Machine, re-imagined as a terminal TUI on modern foundation models. Auto-detects CUDA / MPS / MLX / ROCm / DirectML / XPU / CPU."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "WireML Contributors" }]
keywords = ["teachable-machine", "tui", "clip", "classification", "transfer-learning"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "textual>=0.85.0",
    "rich>=13.9.4",
    "typer>=0.15.1",
    "pydantic>=2.10.4",
    "numpy>=1.26,<2.3",
    "pillow>=11.0.0",
]

[project.optional-dependencies]
ml = [
    "torch>=2.4.0",
    "torchvision>=0.19.0",
    "transformers>=4.46.0",
    "accelerate>=1.2.0",
    "onnxruntime>=1.20.0",
]
mlx = ["mlx>=0.20.0"]
directml = ["onnxruntime-directml>=1.20.0"]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "textual-dev>=1.7.0",
    "ruff>=0.8.4",
    "mypy>=1.14.0",
]

[project.scripts]
wireml = "wireml.cli:app"

[project.urls]
Homepage = "https://github.com/tejasnaladala/wireml"
Repository = "https://github.com/tejasnaladala/wireml"
Issues = "https://github.com/tejasnaladala/wireml/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["wireml"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4", "SIM", "RET"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["B011"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra --strict-markers"
